rabotay

+++ include/picosha2.hpp
@@ -35,7 +35,7 @@ Copyright (C) 2017 okdshin
 #include <string>
 
 namespace picosha2 {
-typedef int32_t word_t;
+typedef int64_t word_t;
 typedef int8_t byte_t;
 
 //static const size_t k_digest_size = 32;
